requirements.txt
notebook
scikit-learn
pandas
matplotlib
seaborn
tensorflow
streamlit
streamlit-option-menu
sklearn-som
plotly